git-svn-id: http://cscdev.lvc.edu/svn/c331.2007/wmb001/CSC500@3385 80dbfefb-8c69-48b5-a6b0-056a8968c603

+++ ResolverTester/test_h_file.h
@@ -15,7 +15,7 @@ namespace {
 
 		void func1();
 
-		int func2();
+		double func2();
 
 		double func3(double i);
 	};
